+++ src/main/java/org/springframework/beans/BeansException.java
@@ -1,0 +1,16 @@
+package org.springframework.beans;
+
+/**
+ * @author derekyi
+ */
+public class BeansException extends RuntimeException {
+
+    public BeansException(String msg) {
+        super(msg);
+    }
+
+    public BeansException(String msg, Throwable cause) {
+        super(msg, cause);
+    }
+
+}