+++ src/main/java/org/springframework/beans/factory/support/AbstractAutowireCapableBeanFactory.java
@@ -1,0 +1,28 @@
+package org.springframework.beans.factory.support;
+
+import org.springframework.beans.BeansException;
+import org.springframework.beans.factory.config.BeanDefinition;
+
+/**
+ * @author derekyi
+ */
+public abstract class AbstractAutowireCapableBeanFactory extends AbstractBeanFactory {
+
+    @Override
+    protected Object createBean(String beanName, BeanDefinition beanDefinition) throws BeansException {
+        return doCreateBean(beanName, beanDefinition);
+    }
+
+    protected Object doCreateBean(String beanName, BeanDefinition beanDefinition) {
+        Class beanClass = beanDefinition.getBeanClass();
+        Object bean;
+        try {
+            bean = beanClass.newInstance();
+        } catch (InstantiationException | IllegalAccessException e) {
+            throw new BeansException("Instantiation of bean failed", e);
+        }
+
+        registerSingleton(beanName, bean);
+        return bean;
+    }
+}