+++ src/main/java/org/springframework/beans/factory/support/AbstractBeanFactory.java
@@ -1,0 +1,43 @@
+package org.springframework.beans.factory.support;
+
+import java.util.Objects;
+import org.springframework.beans.BeansException;
+import org.springframework.beans.factory.BeanFactory;
+import org.springframework.beans.factory.config.BeanDefinition;
+
+/**
+ * @author derekyi
+ */
+public abstract class AbstractBeanFactory extends DefaultSingletonBeanRegistry implements BeanFactory {
+
+    @Override
+    public Object getBean(String beanName) throws BeansException {
+        Object bean = getSingleton(beanName);
+        if (Objects.nonNull(bean)) {
+            return bean;
+        }
+
+        BeanDefinition beanDefinition = getBeanDefinition(beanName);
+        return createBean(beanName, beanDefinition);
+    }
+
+    /**
+     * 获取beanDefinition
+     *
+     * @param beanName beanName
+     * @return beanDefinition
+     * @throws BeansException BeanDefinition不存在时
+     */
+    protected abstract BeanDefinition getBeanDefinition(String beanName) throws BeansException;
+
+    /**
+     * 创建bean
+     *
+     * @param beanName       beanName
+     * @param beanDefinition beanDefinition
+     * @return bean
+     * @throws BeansException 创建bean失败
+     */
+    protected abstract Object createBean(String beanName, BeanDefinition beanDefinition) throws BeansException;
+
+}