+++ src/main/java/org/springframework/beans/factory/support/DefaultSingletonBeanRegistry.java
@@ -1,0 +1,24 @@
+package org.springframework.beans.factory.support;
+
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+import org.springframework.beans.factory.config.SingletonBeanRegistry;
+
+/**
+ * @author derekyi
+ */
+public class DefaultSingletonBeanRegistry implements SingletonBeanRegistry {
+
+    private final Map<String, Object> singletonObjects = new ConcurrentHashMap<>();
+
+    @Override
+    public void registerSingleton(String beanName, Object singletonObject) {
+        singletonObjects.put(beanName, singletonObject);
+    }
+
+    @Override
+    public Object getSingleton(String beanName) {
+        return singletonObjects.get(beanName);
+    }
+
+}